Make all generated headers fully C compatible.
Later headers are already C compatible; extend this compatibility to
earlier headers too. This will make it possible to use these headers
with a C or C++ compiler.

The only change is really changing 'struct X' to 'typedef struct X ...
X' throughout. IDA already implicitly adds this kind of typedef (which
is why IDA was ok with the old headers).